+++ .eslintrc.js
@@ -1,11 +1,3 @@
-const cfgExtends = [
-  'eslint:recommended',
-  'plugin:@typescript-eslint/strict-type-checked',
-]
-if (process.env.NODE_ENV === 'stage') {
-  cfgExtends.push('plugin:@typescript-eslint/stylistic-type-checked')
-}
-
 /** @type {import('eslint').ESLint.ConfigData} */
 module.exports = {
   root: true,
@@ -16,7 +8,14 @@ module.exports = {
     commonjs: true,
   },
   plugins: ['@typescript-eslint'],
-  extends: cfgExtends,
+  extends: [
+    'eslint:recommended',
+    'plugin:@typescript-eslint/strict-type-checked',
+  ].concat(
+    process.env.NODE_ENV === 'stage'
+      ? 'plugin:@typescript-eslint/stylistic-type-checked'
+      : [],
+  ),
   parserOptions: {
     project: true,
   },