+++ src/test/suite-local/caseTransform.test.ts
@@ -1,72 +1,32 @@
 import { strict } from 'assert'
 import { describe, it } from 'mocha'
-import type { WordCase } from '../../utils/caseTransformHelper'
+import type { ComplexWordCase, WordCase } from '../../utils/caseTransformHelper'
 import {
   caseTransform,
-  getWordsByCase,
+  // getWordsByCase,
   joinCaseActions,
-  switchWordCase,
+  // switchWordCase,
 } from '../../utils/caseTransformHelper'
 
-function testAll(words: string[], casedMap: Record<WordCase, string>) {
-  describe('joinCaseActions', function () {
-    it('should match casedMap', function () {
-      casesList.forEach((wc) => {
-        strict.equal(casedMap[wc], joinCaseActions[wc](words))
-      })
-    })
-  })
-
-  describe(`#${getWordsByCase.name}()`, function () {
-    it('should deepEqual words', function () {
-      casesList.forEach((wc) => {
-        if (wc === 'lowerCase' || wc === 'upperCase') {
-          strict.deepEqual(getWordsByCase(casedMap[wc], wc), [
-            casedMap.lowerCase,
-          ])
-        } else {
-          strict.deepEqual(getWordsByCase(casedMap[wc], wc), words)
-        }
-      })
-    })
-  })
-
-  describe('casesReMap', function () {
-    it('should match casedMap', function () {
-      casesList.forEach((wc) => {
-        strict.ok(casesReMap[wc].test(casedMap[wc]))
-      })
-    })
-  })
-
-  describe(`#${switchWordCase.name}()`, function () {
-    it('should return expect word case', function () {
-      casesList.forEach((wc) => {
-        if (wc === 'lowerCase' || wc === 'upperCase') {
-          return
-        }
-        strict.equal(switchWordCase(casedMap[wc]), wc)
-      })
-    })
-  })
-
-  describe(`#${caseTransform.name}()`, function () {
-    it('should return expected cased word', function () {
-      casesList.forEach((wc) => {
-        if (wc === 'lowerCase' || wc === 'upperCase') {
-          return
-        }
-        casesList.forEach((wc2) => {
-          strict.equal(caseTransform(casedMap[wc], wc2), casedMap[wc2])
-        })
-      })
-    })
-  })
+const casedMap: Record<WordCase, string> = {
+  lowerCase: 'itisagoodweather',
+  upperCase: 'ITISAGOODWEATHER',
+  titleCase: 'It Is A Good Weather',
+  noCase: 'it is a good weather',
+  kebabCase: 'it-is-a-good-weather',
+  pascalCase: 'ItIsAGoodWeather',
+  sentenceCase: 'It is a good weather',
+  snakeCase: 'it_is_a_good_weather',
+  pathCase: 'it/is/a/good/weather',
+  dotCase: 'it.is.a.good.weather',
+  camelCase: 'itIsAGoodWeather',
+  headerCase: 'It-Is-A-Good-Weather',
+  constantCase: 'IT_IS_A_GOOD_WEATHER',
 }
 
-const casesReMap: Record<WordCase, RegExp> = {
-  lowerCase: /^[a-z]+$/,
-  upperCase: /^[A-Z]+$/,
+const words = ['it', 'is', 'a', 'good', 'weather']
+
+const casesReMap: Record<ComplexWordCase, RegExp> = {
   noCase: /^(?:(?<!^) [a-z]+|[a-z]+)+$/,
   sentenceCase: /^([A-Z][a-z]*)(?: [a-z]+)*$/,
   dotCase: /^(?:(?<!^)\.[a-z]+|[a-z]+)+$/,
@@ -79,44 +39,31 @@ const casesReMap: Record<WordCase, RegExp> = {
   kebabCase: /^(?:(?<!^)-[a-z]+|[a-z]+)+$/,
   pascalCase: /^(?:[A-Z][a-z]*)+$/,
 }
-const casesList = Object.keys(casesReMap) as WordCase[]
 
-describe('Low Level', function () {
-  const words = ['hello', 'world']
-  const casedMap: Record<WordCase, string> = {
-    lowerCase: 'helloworld',
-    upperCase: 'HELLOWORLD',
-    titleCase: 'Hello World',
-    noCase: 'hello world',
-    kebabCase: 'hello-world',
-    pascalCase: 'HelloWorld',
-    sentenceCase: 'Hello world',
-    snakeCase: 'hello_world',
-    dotCase: 'hello.world',
-    camelCase: 'helloWorld',
-    constantCase: 'HELLO_WORLD',
-    pathCase: 'hello/world',
-    headerCase: 'Hello-World',
-  }
-  testAll(words, casedMap)
+const complexCasesList = Object.keys(casesReMap) as ComplexWordCase[]
+
+describe('casesReMap', function () {
+  it('should match casedMap', function () {
+    complexCasesList.forEach((wc) => {
+      strict.ok(casesReMap[wc].test(casedMap[wc]))
+    })
+  })
 })
 
-describe('High Level', function () {
-  const casedMap: Record<WordCase, string> = {
-    lowerCase: 'itisagoodweather',
-    upperCase: 'ITISAGOODWEATHER',
-    titleCase: 'It Is A Good Weather',
-    noCase: 'it is a good weather',
-    kebabCase: 'it-is-a-good-weather',
-    pascalCase: 'ItIsAGoodWeather',
-    sentenceCase: 'It is a good weather',
-    snakeCase: 'it_is_a_good_weather',
-    pathCase: 'it/is/a/good/weather',
-    dotCase: 'it.is.a.good.weather',
-    camelCase: 'itIsAGoodWeather',
-    headerCase: 'It-Is-A-Good-Weather',
-    constantCase: 'IT_IS_A_GOOD_WEATHER',
-  }
-  const words = ['it', 'is', 'a', 'good', 'weather']
-  testAll(words, casedMap)
+describe('joinCaseActions', function () {
+  it('should match casedMap', function () {
+    complexCasesList.forEach((wc) => {
+      strict.equal(casedMap[wc], joinCaseActions[wc](words))
+    })
+  })
+})
+
+describe(`#${caseTransform.name}()`, function () {
+  it('should return expected cased word', function () {
+    complexCasesList.forEach((wc) => {
+      complexCasesList.forEach((wc2) => {
+        strict.equal(caseTransform(casedMap[wc], wc2), casedMap[wc2])
+      })
+    })
+  })
 })