+++ src/utils/caseTransformHelper.ts
@@ -1,9 +1,27 @@
+export type WordCase =
+  | 'lowerCase'
+  | 'upperCase'
+  | 'dotCase'
+  | 'pathCase'
+  | 'snakeCase'
+  | 'kebabCase'
+  | 'noCase'
+  | 'sentenceCase'
+  | 'constantCase'
+  | 'pascalCase'
+  | 'camelCase'
+  | 'titleCase'
+  | 'headerCase'
+
+export type ComplexWordCase = Exclude<WordCase, 'lowerCase' | 'upperCase'>
+
 /**
  * Functions join the `words` (raw`[a-zA-Z_-$][\w$]*`) array and transform to target case.
  */
-export const joinCaseActions = {
-  lowerCase: (words: string[]) => words.join(''),
-  upperCase: (words: string[]) => words.join('').toUpperCase(),
+export const joinCaseActions: Record<
+  ComplexWordCase,
+  (words: string[]) => string
+> = {
   dotCase: (words: string[]) => words.join('.'),
   pathCase: (words: string[]) => words.join('/'),
   snakeCase: (words: string[]) => words.join('_'),
@@ -38,182 +56,52 @@ export const joinCaseActions = {
     ),
 }
 
-export type WordCase = keyof typeof joinCaseActions
-export const casesList = Object.keys(joinCaseActions) as WordCase[]
-
-/**
- * Switch the sequence case from raw`[a-zA-Z_$][\w$]*`
- * @param sequence a sequence making sequence case
- * @returns {WordCase} case
- */
-export function switchWordCase(sequence: string): WordCase {
-  const reLowerLetter = /[a-z]/
-  const reUpperLetter = /[A-Z]/
-
-  if (reLowerLetter.test(sequence[0])) {
-    const reCaseSplitter = /( )|(_)|(-)|(\.)|(\/)/
-    const matches = sequence.match(reCaseSplitter)
-
-    if (!matches) {
-      const reCamelCase = /^([a-z]+)(?:[A-Z][a-z]*)*$/
-      if (reCamelCase.test(sequence)) {
-        return 'camelCase'
-      }
-      return 'noCase'
-    }
-
-    const idx = matches.findIndex((v, i) => v && i)
-    switch (idx) {
-      case 1:
-        return 'noCase'
-      case 2:
-        return 'snakeCase'
-      case 3:
-        return 'kebabCase'
-      case 4:
-        return 'dotCase'
-      case 5:
-        return 'pathCase'
-    }
-  } else if (reUpperLetter.test(sequence[0])) {
-    const reCaseSplitter = /( )|(_)|(-)/
-    const matches = sequence.match(reCaseSplitter)
-
-    if (!matches) {
-      const reConstantCase = /^(?:(?<!^)_[A-Z]+|[A-Z]+)+$/
-      if (reConstantCase.test(sequence)) {
-        return 'constantCase'
-      }
-      return 'pascalCase'
-    }
-
-    const idx = matches.findIndex((v, i) => v && i)
-    switch (idx) {
-      case 1:
-        if (reLowerLetter.test(sequence.split(' ', 2)[1][0])) {
-          return 'sentenceCase'
-        }
-        return 'titleCase'
-      case 2:
-        return 'constantCase'
-      case 3:
-        return 'headerCase'
-    }
-  }
-
-  return 'lowerCase'
-}
-
-/**
- * RegExp(s) for getting words from selected case.
- */
-const reGetWordsByCase: Record<
-  Exclude<WordCase, 'lowerCase' | 'upperCase'>,
-  RegExp | RegExp[]
-> = (function () {
-  const reAllLowerCase = /[a-z]+/g
-  const reAllCapitalCase = /[A-Z][a-z]*/g
-  return {
-    noCase: reAllLowerCase,
-    snakeCase: reAllLowerCase,
-    kebabCase: reAllLowerCase,
-    pathCase: reAllLowerCase,
-    dotCase: reAllLowerCase,
-    constantCase: /[A-Z]+/g,
-    titleCase: reAllCapitalCase,
-    headerCase: reAllCapitalCase,
-    pascalCase: reAllCapitalCase,
-    // 0=1; 1=0
-    sentenceCase: [/^([A-Z][a-z]*)(?: [a-z]+)*$/, /(?<= )[a-z]+/g],
-    camelCase: [/^([a-z]+)(?:[A-Z][a-z]*)*$/, reAllCapitalCase],
-  }
-})()
-
-/**
- * Transform one sequence from `wordCase` to `words` array.
- * @param wc word case
- * @returns transformed sequence
- */
-export function getWordsByCase(sequence: string, wc: WordCase) {
-  switch (wc) {
-    case 'lowerCase':
-      return [sequence]
-    case 'upperCase':
-      return [sequence.toLowerCase()]
-    default: {
-      const re = reGetWordsByCase[wc]
-      if (re instanceof RegExp) {
-        const matches = sequence.match(re)
-        if (!matches) {
-          return null
-        }
-        switch (wc) {
-          case 'titleCase':
-          case 'constantCase':
-          case 'pascalCase':
-          case 'headerCase':
-            return matches.map((w) => w.toLowerCase())
-          // case 'dotCase':
-          // case 'pathCase':
-          // case 'snakeCase':
-          // case 'paramCase':
-          // case 'noCase':
-          default:
-            return matches
-        }
-      } else {
-        const matchHead = sequence.match(re[0])
-        if (!matchHead) {
-          return null
-        }
-        switch (wc) {
-          case 'camelCase':
-            return [matchHead[1]].concat(
-              sequence.match(re[1])?.map((w) => w.toLowerCase()) ?? [],
-            )
-          case 'sentenceCase':
-            return [matchHead[1].toLowerCase()].concat(
-              sequence.match(re[1]) ?? [],
-            )
-          default:
-            return null
-        }
-      }
-    }
-  }
-}
-
 /**
  * Only transform a sequence string belongs {@link WordCase}.
  * @param sequence sequence to be transform
- * @param targetCase transform target case
+ * @param targetWc transform target case
  * @returns transformed sequence
  */
-export function caseTransform(sequence: string, targetCase: WordCase) {
-  switch (targetCase) {
+export function caseTransform(sequence: string, targetWc: WordCase) {
+  switch (targetWc) {
     case 'lowerCase':
       return sequence.toLowerCase()
     case 'upperCase':
       return sequence.toUpperCase()
     default: {
-      const reSpiltChar = /[\d$]|$/g
-      const targets: string[] = []
+      const reSpiltChar = /[^a-zA-Z-/_.\s]+|$/g
+      const targetSegments: string[] = []
 
       let lastIdx = 0
       for (const { index, 0: spiltChar } of sequence.matchAll(reSpiltChar)) {
         const piece = sequence.slice(lastIdx, index)
         if (piece.length) {
-          const wc = switchWordCase(piece)
-          const words = getWordsByCase(piece, wc)
+          const words = getWords(piece)
           if (words) {
-            targets.push(joinCaseActions[targetCase](words))
+            targetSegments.push(joinCaseActions[targetWc](words))
           }
         }
         lastIdx = index! + spiltChar.length
-        targets.push(spiltChar)
+        targetSegments.push(spiltChar)
       }
 
-      return targets.join('')
+      return targetSegments.join('')
     }
   }
 }
+
+export function getWords(text: string) {
+  const words = text.split(/[-/_.\s]+/)
+  const res: string[] = []
+  for (const word of words) {
+    for (const matches of word.matchAll(getWords.reGetWords)) {
+      if (matches[1]) {
+        res.push(matches[1].toLowerCase(), matches[2].toLowerCase())
+      } else {
+        res.push(matches[0].toLowerCase())
+      }
+    }
+  }
+  return res.length ? res : null
+}
+getWords.reGetWords = /([A-Z]+)([A-Z][a-z]+)|[A-Z][a-z]+|[a-z]+|[A-Z]+/g